Move some stuff around, add fallback linkedin behavior for cases where we dont have verifiable credentials
